Improve speed for large coverage reports and source directories
No longer look at all the directories, but just return the first
that matches all the files in the coverage report. If no matching
source dirs are found, exit with status code 1.

This is different behaviour than before, hence the minor version
bump. (Should have technically been a major version bump, but the
difference in behaviour isn't that huge).